[ipq806x] qca-nss-drv : CR Fixed:533303
When a Scattered SKB fragment is given to NetAP, in the
HLOS descriptor the payload offset field was set to
frag->offset. However the descriptor buffer already  points
to payload_offset(in the page) when mapped using skb_frag_dma_map.
Hence the desc->payload_offset shall be sent as zero to NetAP so that
NetAp starts reading the packet from correct offset.This was causing
LUCI Application to read wrong data when TSO/SG is enabled

Change-Id: I044bb3695e96cb631bb3d2f6e4684d331a25d26d
Signed-off-by: Radhakrishna Jiguru <rjiguru@codeaurora.org>
Reviewed-by: Abhishek Rastogi <arastogi@codeaurora.org>