Fix bridge exception path crash with 6RD
CRs-Fixed: 574363

Summary: When the link goes down and 6rd rules are flushed, there is a
chance that some exception packets that reach host have been partially
processed in the fast path by NSS, and might not have a MAC header.
Hence check for the presense of a MAC header when accessing the MAC
addreses.

Change-Id: Id093821368a05701e4ea0b1dbebb4e1174b583d2
Signed-off-by: Kiran Kumar C.S.K <kkumarcs@codeaurora.org>