[qca-nss-drv]: Add Changes needed for jumbo support
CR Fixed:555993
The Jumbo Support in the system needs the allocation of memory buffers of sizes > 9600 Bytes.
Added the logic to dynamically change the buffer allocation sizes based on the Maximu MTU value
configured in the system.

The change for jumbo support is needed in NetAP firmware,qca-nss-drv and qca-nss-gmac driver.This commit handles
the needed changes in NSS driver
Signed-off-by: Radhakrishna Jiguru <rjiguru@codeaurora.org>

Change-Id: Icf4bca6b81aac21444917b52158924cb9404661b
Signed-off-by: Radhakrishna Jiguru <rjiguru@codeaurora.org>