Disable frequency statistics from NSS when auto-scaling is turned off
CRs-Fixed: 592908

Note: This change is part of a list of changes being done to the NSS
to improve small packet performance and consistency. Disabling frequency
statistics when auto scaling is turned off ensure we do not allocate buffers
(1000 per second) in NSS when these syncs are not necesary at host, and also
reduce CPU utilization in NSS required for processing these syncs to host.

Change-Id: I64d65c0adb8086b6af6e2dea40e5e88b504a9def
Signed-off-by: Kiran Kumar C.S.K <kkumarcs@codeaurora.org>