[qca-nss-drv]: Enable TSO logic for non scatttered frames
CRs-Fixed: 608375,61184
When CONFIG_HIGHMEM kernel config option is enabled,on DB149 boards
Linux is sending a single larger buffer in non scattered manner.

The TSO offload implementation assumes that the larger chunks will
be given in scatteres.Even ethtool says it is required  to enable
SG support for enabling GSO/TSO support .

This commit addresses the needed changes in host to enable TSO support
in NetAP when non scattered larger chunk of packet is given to it.

Change-Id: If3cae25df4118200d1b1a56cdae13dad14b5ff8b
Signed-off-by: Radha krishna Simha Jiguru <rjiguru@codeaurora.org>