[nss_core] Workaround for scatter gather issue in the NSS
Currently we do not handle scatter gather in a way that is robust in
all situations, in the NSS. This results in crashes when scattered
packets are bounced for shaping.

This fix provides a temporary workaround without impacting the SAMBA
performance. However, this is not a long term solution and has its own
risk when used. These are detailed in the comments.

CRs-Fixed: 658965

Change-Id: I26cfdcd394dda4c047311d387548653e184b66f8
Signed-off-by: Sakthi Vignesh Radhakrishnan <sradhakr@codeaurora.org>