[nss_qdisc] Fix packet corruption in the bounce queue
The bounce callback function performed an un-protected enqueue to a queue
that is lock protected. This results in occational packet corruption which
leads to a crash.

Additional fix: When bounce enqueue fails, the packet must be discarded
and not transmitted. Such transmitted packets will skip bridge shaping
which is wrong.

CRs-Fixed: 659001

Change-Id: I73d56e23b94e1d18776c61f79ba02ca1aa341429
Signed-off-by: Sakthi Vignesh Radhakrishnan <sradhakr@codeaurora.org>