[qca-nss-drv] Block grafting of nssbf and nsswrr/nsswfq root classes
Root classes in nssbf and nsswrr/nsswfq qdiscs are not meant to be
used. The ideal solution is to remove them, but since there is
StreamBoost dependency on them right now, that can't be done. So
will block attempts to graft them.

Populate set_default field for nsspfifo and nsscodel qdisc dump.

Change-Id: Ibccc7fddd3277d907e884e74989614be9e203e73
CRs-Fixed: 681445,681648
Signed-off-by: Mark Kofman <mkofman@codeaurora.org>