[qca-nss-qdisc] Fix bug in accounting for drops due to h2n failure
The qdisc proxy did not account for drops that happened due to a
failure in sending packets down to the nss. This situation could
arise when the h2n buffer ring is full.

This fix accounts for the drops by properly matching the skb->priority
with a qos_tag and updating drops bottom up.

CRs-Fixed: 701486
Change-Id: I88258bd1e57e8299e36aa80202b43b74b3f4cfc3
Signed-off-by: Sakthi Vignesh Radhakrishnan <sradhakr@codeaurora.org>