Fix issue with forcing link speed/duplex for SGMII and QSGMII interfaces
CRs-Fixed: 771778

Description:

1.) We have new SGMII configurations on AP160.2 where link polling
may be disabled or enabled, but link is always forced to certain speeds. The
GMAC driver currently relies on link polling flag to detect and set speeds for
SGMII. We need to fix this to use 'forced_speed' values instead.

2.) Removed redundant link status read from phy in nss_gmac_check_link(), since
it is breaking link status check for non-generic PHYs (AQ PHY).

3.) Clean up to remove redundant macros for speed/duplex and use only ethtool.h macros
4.) Clean up of function nss_gmac_check_phy_init() to handle all cases

Change-Id: I18f69bffe81e2d3b1c8e92ae43df18eaff62b66c
Signed-off-by: Kiran Kumar C.S.K <kkumarcs@codeaurora.org>