[qca-nss-qdisc] Handle NULL opt
New version of TC provides a NULL netlink message when users dont configure
any of the supported tunable parameters. Therefore a create request with NULL
message is still valid and has to be processed.

This affects qdiscs that only have non-mandatory input parameters. In our
list of supported qdiscs nssbf, nsshtb and nssprio falls in this category.

Change-Id: Ic303eeea38fca0f0af25de8a7fee66d89c9d2b61
Signed-off-by: Sakthi Vignesh Radhakrishnan <sradhakr@codeaurora.org>