[qca-nss-drv] Added fabric scaling to coconut
Added fabric scaling 0 and 1 to coconut. Since we are on coconut,

We will check in the DT_SUPPORT. If DT_SUPPORT is enabled, then we will
attempt to get fabric pointers to the handlers. If not sucessful, then we
will continue. Requests done to firmware will check for NULL before issuing
requests

Change-Id: Ia82afacd3a792b4bbbbb1b8d24f7ea28d908684f
Signed-off-by: Thomas Wu <wthomas@codeaurora.org>