[nss-host] qca-nss-ecm: Use another path check for bridge acceleration
The conntrack notifier was moved from frontend directories
to the ECM's top directory. Conntrack notifier is
common for the frontends anymore. So, instead of checking
the conntrack notifier directory, now we check the frontend
directories (ecm_nss_ipv4 or ecm_sfe_ipv4) to decide if the
underlying acceleration engine is NSS or SFE.

Change-Id: If1f24e5c06675504af25ff5bcaf77553abd2b202
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>