[qca-nss-clients] fix for IPsec issue
The patch fixes an issue where we would try to perform
a route lookup for identifying the associated NETDEV
for the IPsec. This would break in cases where the
exception happens and the packet is encapsulated with ESP.
The outer headers will be potentially point to actual
physical device instead.

Change-Id: I2ccf36e4daf503406526b6091564d80b01e95ef5
Signed-off-by: Samarjeet Banerjee <banerjee@codeaurora.org>